refactor: split out XMLHashOptions to own file
