+++ Source/XMLHash.swift
@@ -36,32 +36,6 @@ import FoundationXML
 
 let rootElementName = "SWXMLHash_Root_Element"
 
-/// Parser options
-public class XMLHashOptions {
-    internal init() {}
-
-    /// determines whether to parse the XML with lazy parsing or not
-    public var shouldProcessLazily = false
-
-    /// determines whether to parse XML namespaces or not (forwards to
-    /// `XMLParser.shouldProcessNamespaces`)
-    public var shouldProcessNamespaces = false
-
-    /// Matching element names, element values, attribute names, attribute values
-    /// will be case insensitive. This will not affect parsing (data does not change)
-    public var caseInsensitive = false
-
-    /// Encoding used for XML parsing. Default is set to UTF8
-    public var encoding = String.Encoding.utf8
-
-    /// Any contextual information set by the user for encoding
-    public var userInfo = [CodingUserInfoKey: Any]()
-
-    /// Detect XML parsing errors... defaults to false as this library will
-    /// attempt to handle HTML which isn't always XML-compatible
-    public var detectParsingErrors = false
-}
-
 /// Simple XML parser
 public class XMLHash {
     let options: XMLHashOptions
